+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 - Added `nstd_core_cstr_as_slice`.
 - Moved the `cstr` module out of `str`.
 ### `nstd.os`
-- Added more heap control in `windows.alloc`.
+- Added `windows.alloc.heap` module.
 - Added `windows.def` module.
 - Added `windows.io` module.
 - Added `linux.alloc` module.